# backend/swagger.yaml
openapi: 3.0.0
info:
  title: PMF Checker API
  version: 1.0.0
  description: API documentation for PMF Checker Backend - Browser-Use task automation platform

servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.pmfchecker.com
    description: Production server

tags:
  - name: Auth
    description: Authentication endpoints
  - name: Tasks
    description: Browser automation task management
  - name: Webhooks
    description: Browser-Use Cloud webhooks

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
          example: "507f1f77bcf86cd799439011"
        email:
          type: string
          format: email
          example: "user@example.com"
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."

    Task:
      type: object
      properties:
        _id:
          type: string
          example: "507f1f77bcf86cd799439011"
        user:
          type: string
          example: "507f1f77bcf86cd799439011"
        browserUseId:
          type: string
          example: "902de05c-742a-4a78-afcc-28e2f7f28d95"
        prompt:
          type: string
          example: "Go to LinkedIn and find 10 software development jobs"
        status:
          type: string
          enum: [STARTED, RUNNING, DONE, ERROR, FINALIZING]
          example: "DONE"
        result:
          type: object
          properties:
            output:
              type: string
            claim:
              type: string
            pricingVisible:
              type: boolean
        logs:
          type: array
          items:
            type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    Error:
      type: object
      properties:
        error:
          type: string
          example: "Invalid credentials"

paths:
  /api/auth/signup:
    post:
      summary: Register a new user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                password:
                  type: string
                  format: password
                  minLength: 6
                  example: "password123"
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: User already exists or invalid data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/auth/login:
    post:
      summary: Login a user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                password:
                  type: string
                  format: password
                  example: "password123"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/tasks:
    post:
      summary: Create a new browser automation task
      tags: [Tasks]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - taskPrompt
              properties:
                taskPrompt:
                  type: string
                  example: "Go to LinkedIn and find 10 software development jobs"
                  description: Natural language description of the task to automate
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  taskId:
                    type: string
                    example: "902de05c-742a-4a78-afcc-28e2f7f28d95"
                    description: Browser-Use Cloud task ID
                  dbId:
                    type: string
                    example: "507f1f77bcf86cd799439011"
                    description: MongoDB document ID
                  status:
                    type: string
                    example: "STARTED"
        '400':
          description: Bad request - taskPrompt missing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      summary: Get all tasks for authenticated user
      tags: [Tasks]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/tasks/{id}:
    get:
      summary: Get a task by ID (MongoDB ID or Browser-Use ID)
      tags: [Tasks]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Task ID (MongoDB _id or browserUseId)
          example: "902de05c-742a-4a78-afcc-28e2f7f28d95"
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/tasks/{taskId}/stream:
    get:
      summary: Server-Sent Events stream for real-time task updates
      tags: [Tasks]
      description: |
        Opens a Server-Sent Events (SSE) connection for real-time task updates.
        Connect using EventSource in JavaScript:
        ```javascript
        const eventSource = new EventSource('/api/tasks/abc123/stream');
        eventSource.onmessage = (event) => {
          const { type, data } = JSON.parse(event.data);
          console.log(type, data);
        };
        ```
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: string
          description: Browser-Use task ID
          example: "902de05c-742a-4a78-afcc-28e2f7f28d95"
      responses:
        '200':
          description: SSE stream established
          content:
            text/event-stream:
              schema:
                type: string
                example: |
                  data: {"type":"status","data":{"step":"RUNNING","message":"Processing..."}}
                  
                  data: {"type":"done","data":{"output":"Task completed"}}

  /api/webhooks/browser-use:
    post:
      summary: Webhook endpoint for Browser-Use Cloud updates
      tags: [Webhooks]
      description: Receives task status updates from Browser-Use Cloud API
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  example: "agent.task.status_update"
                timestamp:
                  type: string
                  format: date-time
                payload:
                  type: object
                  properties:
                    task_id:
                      type: string
                    status:
                      type: string
                      enum: [started, processing, finished, failed]
      responses:
        '200':
          description: Webhook processed successfully
        '400':
          description: Invalid webhook payload
        '401':
          description: Invalid webhook secret

security:
  - bearerAuth: []